+++ Battery-Control-By-Reinforcement-Learning/weather_data_bid.csv
@@ -1,0 +1,49 @@
+,year,month,day,hour,Pressure,temperature,u-component of wind,v-component of wind,Relative humidity,Total cloud cover,Total precipitation,radiation flux
+0,2023,1,31,0.0,1012.3397216796875,1.3651214599609602,2.5789146423339844,-0.7059659957885742,52.996826171875,0.0,0.0,0.0
+1,2023,1,31,0.5,1012.7173156738281,1.1310058593750227,2.2299914360046387,-0.8267483711242676,54.77342987060547,0.0,0.0,0.0
+2,2023,1,31,1.0,1013.0949096679688,0.8968902587890852,1.881068229675293,-0.9475307464599609,56.55003356933594,0.0,0.0,0.0
+3,2023,1,31,1.5,1013.4716186523438,0.7374435424804915,1.8001627922058105,-0.9570932388305664,57.06071758270264,0.0,0.0,0.0
+4,2023,1,31,2.0,1013.8483276367188,0.5779968261718977,1.7192573547363281,-0.9666557312011719,57.571401596069336,0.0,0.0,0.0
+5,2023,1,31,2.5,1014.1520690917969,-0.04807128906247726,1.1333837509155273,-1.1361198425292969,63.916874408721924,0.244140625,0.0,0.0
+6,2023,1,31,3.0,1014.455810546875,-0.6741394042968523,0.5475101470947266,-1.3055839538574219,70.26234722137451,0.48828125,0.0,0.0
+7,2023,1,31,3.5,1014.8145141601562,-0.7993072509765398,0.5382390022277832,-1.0807199478149414,67.9108943939209,3.3935546875,0.0,0.0
+8,2023,1,31,4.0,1015.1732177734375,-0.9244750976562273,0.5289678573608398,-0.8558559417724609,65.55944156646729,6.298828125,0.0,0.0
+9,2023,1,31,4.5,1015.5029602050781,-0.8883575439452898,0.728121280670166,-0.7097616195678711,63.87635087966919,7.470703125,0.0,0.0
+10,2023,1,31,5.0,1015.8327026367188,-0.8522399902343523,0.9272747039794922,-0.5636672973632812,62.193260192871094,8.642578125,0.0,0.0
+11,2023,1,31,5.5,1016.196044921875,-0.9491180419921648,0.9819092750549316,-0.47583580017089844,61.88687229156494,6.2744140625,0.0,0.0
+12,2023,1,31,6.0,1016.5593872070312,-1.0459960937499773,1.036543846130371,-0.3880043029785156,61.58048439025879,3.90625,0.0,0.0
+13,2023,1,31,6.5,1016.9088134765625,-1.2049087524413835,1.1233725547790527,-0.23424434661865234,61.933836936950684,2.2216796875,0.0,1.078125
+14,2023,1,31,7.0,1017.2582397460938,-1.3638214111327898,1.2102012634277344,-0.08048439025878906,62.28718948364258,0.537109375,0.0,2.15625
+15,2023,1,31,7.5,1017.7081604003906,-0.852232360839821,0.8876833915710449,0.10818004608154297,62.27555179595947,0.2685546875,0.0,52.453125
+16,2023,1,31,8.0,1018.1580810546875,-0.34064331054685226,0.5651655197143555,0.296844482421875,62.26391410827637,0.0,0.0,102.75
+17,2023,1,31,8.5,1018.495361328125,0.45302734375002274,1.0404772758483887,-0.18189144134521484,56.99682569503784,0.0,0.0,195.63280367851257
+18,2023,1,31,9.0,1018.8326416015625,1.2466979980468977,1.5157890319824219,-0.6606273651123047,51.729737281799316,0.0,0.0,288.51560735702515
+19,2023,1,31,9.5,1018.906982421875,1.9851165771484602,1.5328941345214844,-0.7280502319335938,50.27046203613281,0.0,0.0,371.2845103740692
+20,2023,1,31,10.0,1018.9813232421875,2.7235351562500227,1.5499992370605469,-0.7954730987548828,48.81118679046631,0.0,0.0,454.0534133911133
+21,2023,1,31,10.5,1018.7911987304688,3.326936340332054,1.4554328918457031,-0.8101787567138672,46.58056020736694,0.0,0.0,512.9585971832275
+22,2023,1,31,11.0,1018.60107421875,3.9303375244140852,1.3608665466308594,-0.8248844146728516,44.34993362426758,0.0,0.0,571.8637809753418
+23,2023,1,31,11.5,1018.1738586425781,4.354959106445335,1.3162531852722168,-0.7727766036987305,43.095420837402344,0.0,0.0,600.6763181686401
+24,2023,1,31,12.0,1017.7466430664062,4.779580688476585,1.2716398239135742,-0.7206687927246094,41.84090805053711,0.0,0.0,629.4888553619385
+25,2023,1,31,12.5,1017.3959350585938,5.133935546875023,1.2096514701843262,-0.8046989440917969,40.5999059677124,0.0,0.0,625.2176151275635
+26,2023,1,31,13.0,1017.0452270507812,5.48829040527346,1.1476631164550781,-0.8887290954589844,39.358903884887695,0.0,0.0,620.9463748931885
+27,2023,1,31,13.5,1016.9884948730469,5.718301391601585,1.0205183029174805,-0.9263086318969727,38.14169979095459,0.0244140625,0.0,584.2100067138672
+28,2023,1,31,14.0,1016.9317626953125,5.94831237792971,0.8933734893798828,-0.9638881683349609,36.924495697021484,0.048828125,0.0,547.4736385345459
+29,2023,1,31,14.5,1016.9998474121094,6.035150146484398,0.7192230224609375,-0.7987990379333496,36.15997886657715,0.4638671875,0.0,481.76096200942993
+30,2023,1,31,15.0,1017.0679321289062,6.121987915039085,0.5450725555419922,-0.6337099075317383,35.39546203613281,0.87890625,0.0,416.04828548431396
+31,2023,1,31,15.5,1017.2840576171875,6.071557617187523,0.14665699005126953,-0.24889659881591797,34.83445739746094,1.611328125,0.0,328.7468521595001
+32,2023,1,31,16.0,1017.5001831054688,6.02112731933596,-0.2517585754394531,0.13591670989990234,34.27345275878906,2.34375,0.0,241.44541883468628
+33,2023,1,31,16.5,1017.7763977050781,5.5435043334961165,-0.590064525604248,0.4604477882385254,40.641103744506836,2.099609375,0.0,152.28520941734314
+34,2023,1,31,17.0,1018.0526123046875,5.065881347656273,-0.928370475769043,0.7849788665771484,47.00875473022461,1.85546875,0.0,63.125
+35,2023,1,31,17.5,1018.3893127441406,4.3316207885742415,-1.0420866012573242,0.9094834327697754,49.19064807891846,1.171875,0.0,31.5625
+36,2023,1,31,18.0,1018.7260131835938,3.5973602294922102,-1.1558027267456055,1.0339879989624023,51.372541427612305,0.48828125,0.0,0.0
+37,2023,1,31,18.5,1018.9446105957031,3.0654541015625227,-1.1122002601623535,1.0820846557617188,53.500911712646484,1.1474609375,0.0,0.0
+38,2023,1,31,19.0,1019.1632080078125,2.5335479736328352,-1.0685977935791016,1.1301813125610352,55.629281997680664,1.806640625,0.0,0.0
+39,2023,1,31,19.5,1019.176513671875,2.231080627441429,-1.1727285385131836,1.139470100402832,56.58789253234863,2.44140625,0.0,0.0
+40,2023,1,31,20.0,1019.1898193359375,1.9286132812500227,-1.2768592834472656,1.148758888244629,57.5465030670166,3.076171875,0.0,0.0
+41,2023,1,31,20.5,1019.1253662109375,1.716859436035179,-1.303785800933838,1.047579288482666,60.92144298553467,15.2099609375,0.0,0.0
+42,2023,1,31,21.0,1019.0609130859375,1.5051055908203352,-1.3307123184204102,0.9463996887207031,64.29638290405273,27.34375,0.0,0.0
+43,2023,1,31,21.5,1018.9334411621094,1.3715301513672102,-1.0546917915344238,1.0561761856079102,66.52603483200073,38.8916015625,0.0,0.0
+44,2023,1,31,22.0,1018.8059692382812,1.2379547119140852,-0.7786712646484375,1.1659526824951172,68.75568675994873,50.439453125,0.0,0.0
+45,2023,1,31,22.5,1018.5681457519531,1.1896835327148665,-0.7568926811218262,0.6861248016357422,71.74531888961792,57.373046875,0.0,0.0
+46,2023,1,31,23.0,1018.330322265625,1.1414123535156477,-0.7351140975952148,0.2062969207763672,74.73495101928711,64.306640625,0.0,0.0
+47,2023,1,31,23.5,1018.1373901367188,1.1371856689453352,0.06706714630126953,-0.3812832832336426,72.65568161010742,57.958984375,0.0,0.0
