+++ Battery-Control-By-Reinforcement-Learning/weather_data_realtime.csv
@@ -1,0 +1,55 @@
+,year,month,day,hour,Pressure,temperature,u-component of wind,v-component of wind,Relative humidity,Total cloud cover,Total precipitation,radiation flux
+0,2023,1,31,15.0,1013.8885498046875,6.803674316406273,1.0642662048339844,-0.9151725769042969,37.400033950805664,0.0,0.0,242.3516182899475
+1,2023,1,31,15.5,1014.3019104003906,6.829370117187523,1.338636875152588,-0.6142702102661133,36.86249923706055,0.0,0.0,152.92580914497375
+2,2023,1,31,16.0,1014.7152709960938,6.855065917968773,1.6130075454711914,-0.3133678436279297,36.32496452331543,0.0,0.0,63.5
+3,2023,1,31,16.5,1015.0106811523438,6.01111755371096,1.3686699867248535,-0.00482177734375,49.20485496520996,0.537109375,0.0,31.75
+4,2023,1,31,17.0,1015.3060913085938,5.167169189453148,1.1243324279785156,0.3037242889404297,62.08474540710449,1.07421875,0.0,0.0
+5,2023,1,31,17.5,1015.6684875488281,4.578897094726585,1.0883126258850098,0.7917613983154297,61.90422821044922,2.2705078125,0.0,0.0
+6,2023,1,31,18.0,1016.0308837890625,3.9906250000000227,1.052292823791504,1.2797985076904297,61.723711013793945,3.466796875,0.0,0.0
+7,2023,1,31,18.5,1016.2345886230469,4.019982910156273,0.5503458976745605,0.36481714248657227,55.005353927612305,14.9169921875,0.0,0.0
+8,2023,1,31,19.0,1016.4382934570312,4.049340820312523,0.04839897155761719,-0.5501642227172852,48.286996841430664,26.3671875,0.0,0.0
+9,2023,1,31,19.5,1016.52587890625,3.1055694580078352,-0.7344813346862793,-0.7267637252807617,49.574557304382324,31.7626953125,0.0,0.0
+10,2023,1,31,20.0,1016.6134643554688,2.1617980957031477,-1.5173616409301758,-0.9033632278442383,50.862117767333984,37.158203125,0.0,0.0
+11,2023,1,31,20.5,1016.6324157714844,1.8625961303711165,-1.2425994873046875,-0.8075037002563477,53.352216720581055,38.818359375,0.0,0.0
+12,2023,1,31,21.0,1016.6513671875,1.5633941650390852,-0.9678373336791992,-0.711644172668457,55.842315673828125,40.478515625,0.0,0.0
+13,2023,1,31,21.5,1016.6832885742188,1.3448730468750227,-0.6901798248291016,-0.8504385948181152,56.42689275741577,42.6513671875,0.0,0.0
+14,2023,1,31,22.0,1016.7152099609375,1.1263519287109602,-0.4125223159790039,-0.9892330169677734,57.01146984100342,44.82421875,0.0,0.0
+15,2023,1,31,22.5,1016.6508483886719,0.7303100585937727,0.16252660751342773,-0.9210634231567383,58.58719491958618,27.34375,0.0,0.0
+16,2023,1,31,23.0,1016.5864868164062,0.33426818847658524,0.7375755310058594,-0.8528938293457031,60.162919998168945,9.86328125,0.0,0.0
+17,2023,1,31,23.5,1016.4032592773438,-0.1482681274413835,0.6889481544494629,-0.9464006423950195,63.92693328857422,5.2734375,0.0,0.0
+18,2023,2,1,0.0,1016.2200317382812,-0.6308044433593523,0.6403207778930664,-1.039907455444336,67.69094657897949,0.68359375,0.0,0.0
+19,2023,2,1,0.5,1016.2713012695312,-0.848692321777321,0.7609610557556152,-1.1090202331542969,64.60243225097656,0.341796875,0.0,0.0
+20,2023,2,1,1.0,1016.3225708007812,-1.0665802001952898,0.8816013336181641,-1.1781330108642578,61.51391792297363,0.0,0.0,0.0
+21,2023,2,1,1.5,1016.2860107421875,-1.5648330688476335,0.9035258293151855,-0.9730033874511719,66.93508625030518,0.0,0.0,0.0
+22,2023,2,1,2.0,1016.2494506835938,-2.0630859374999773,0.925450325012207,-0.7678737640380859,72.35625457763672,0.0,0.0,0.0
+23,2023,2,1,2.5,1016.0173034667969,-2.2946228027343523,0.9462409019470215,-0.6061892509460449,72.55519771575928,0.0,0.0,0.0
+24,2023,2,1,3.0,1015.78515625,-2.5261596679687273,0.9670314788818359,-0.4445047378540039,72.75414085388184,0.0,0.0,0.0
+25,2023,2,1,3.5,1015.6214904785156,-2.669309997558571,0.9912018775939941,-0.13257837295532227,72.94097900390625,0.0,0.0,0.0
+26,2023,2,1,4.0,1015.4578247070312,-2.8124603271484148,1.0153722763061523,0.17934799194335938,73.12781715393066,0.0,0.0,0.0
+27,2023,2,1,4.5,1015.5491333007812,-2.9281524658202898,1.0908551216125488,0.5137825012207031,72.7412052154541,0.0,0.0,0.0
+28,2023,2,1,5.0,1015.6404418945312,-3.0438446044921648,1.1663379669189453,0.8482170104980469,72.35459327697754,0.0,0.0,0.0
+29,2023,2,1,5.5,1015.7559814453125,-3.332395935058571,1.073857307434082,0.7666988372802734,75.51506900787354,0.0,0.0,0.7916666666666666
+30,2023,2,1,6.0,1015.8715209960938,-3.6209472656249773,0.9813766479492188,0.6851806640625,78.67554473876953,0.0,0.0,1.5833333333333333
+31,2023,2,1,7.0,1015.836669921875,-3.5095275878906023,0.7727851867675781,0.4966754913330078,77.1589126586914,0.0,0.0,2.375
+32,2023,2,1,7.5,1015.87255859375,-2.9701141357421648,0.4165825843811035,0.10314273834228516,73.29800319671631,0.0,0.0,52.8125
+33,2023,2,1,8.0,1015.908447265625,-2.4307006835937273,0.060379981994628906,-0.2903900146484375,69.43709373474121,0.0,0.0,103.25
+34,2023,2,1,8.5,1015.6449890136719,-0.995237731933571,0.19930315017700195,-0.013415336608886719,63.481905937194824,0.0,0.0,194.97494554519653
+35,2023,2,1,9.0,1015.3815307617188,0.44022521972658524,0.338226318359375,0.26355934143066406,57.52671813964844,0.0,0.0,286.69989109039307
+36,2023,2,1,9.5,1014.9128112792969,2.4017730712890852,0.32744884490966797,0.9720673561096191,57.59478282928467,0.0,0.0,369.22719383239746
+37,2023,2,1,10.0,1014.444091796875,4.363320922851585,0.31667137145996094,1.6805753707885742,57.6628475189209,0.0,0.0,451.75449657440186
+38,2023,2,1,10.5,1013.6986999511719,6.0773788452148665,0.94818115234375,2.064652442932129,57.9969425201416,0.0,0.0,510.57943868637085
+39,2023,2,1,11.0,1012.9533081054688,7.791436767578148,1.579690933227539,2.4487295150756836,58.331037521362305,0.0,0.0,569.4043807983398
+40,2023,2,1,11.5,1012.0682678222656,9.120103454589866,2.141116142272949,3.219806671142578,50.764065742492676,25.5126953125,0.0,595.8143873214722
+41,2023,2,1,12.0,1011.1832275390625,10.448770141601585,2.7025413513183594,3.9908838272094727,43.19709396362305,51.025390625,0.0,622.2243938446045
+42,2023,2,1,12.5,1010.4274597167969,10.702996826171898,2.687077045440674,4.4320759773254395,43.078349113464355,46.6796875,0.0,610.2012557983398
+43,2023,2,1,13.0,1009.6716918945312,10.95722351074221,2.6716127395629883,4.873268127441406,42.959604263305664,42.333984375,0.0,598.1781177520752
+44,2023,2,1,13.5,1009.1108093261719,11.168580627441429,2.7523746490478516,4.982536315917969,44.51356601715088,41.5283203125,0.0,562.2086305618286
+45,2023,2,1,14.0,1008.5499267578125,11.379937744140648,2.833136558532715,5.091804504394531,46.067527770996094,40.72265625,0.0,526.239143371582
+46,2023,2,1,14.5,1008.138671875,11.424897766113304,2.8131284713745117,5.118502616882324,48.86157512664795,31.1767578125,0.0,459.6991448402405
+47,2023,2,1,15.0,1007.7274169921875,11.46985778808596,2.7931203842163086,5.145200729370117,51.655622482299805,21.630859375,0.0,393.1591463088989
+48,2023,2,1,15.5,1007.3437194824219,11.309236145019554,2.750004768371582,5.0998029708862305,53.6168270111084,11.0595703125,0.0,314.2778902053833
+49,2023,2,1,16.0,1006.9600219726562,11.148614501953148,2.7068891525268555,5.054405212402344,55.57803153991699,0.48828125,0.0,235.39663410186768
+50,2023,2,1,16.5,1006.5874633789062,10.806214904785179,2.5148239135742188,4.840087890625,58.25648593902588,4.19921875,0.0,149.26081705093384
+51,2023,2,1,17.0,1006.2149047851562,10.46381530761721,2.322758674621582,4.625770568847656,60.934940338134766,7.91015625,0.0,63.125
+52,2023,2,1,17.5,1006.1578674316406,10.389054870605491,2.3591575622558594,4.445526123046875,62.621758460998535,13.0859375,0.0,31.5625
+53,2023,2,1,18.0,1006.100830078125,10.314294433593773,2.3955564498901367,4.265281677246094,64.3085765838623,18.26171875,0.0,0.0
